move table view code into an extension
