+++ amatracker/Percentages View Controller/PercentagesViewController.swift
@@ -2,7 +2,7 @@ import Foundation
 import UIKit
 import Mixpanel
 
-class PercentagesViewController : UIViewController, UITableViewDataSource, UITableViewDelegate  {
+class PercentagesViewController : UIViewController {
     
     let POUNDS_TO_KILOS_RATE : Float = 2.2
     
@@ -15,6 +15,8 @@ class PercentagesViewController : UIViewController, UITableViewDataSource, UITab
     let percentages : [Float] = [0.5, 0.55, 0.6, 0.65,
         0.7, 0.75, 0.8, 0.85,
         0.9, 0.95, 1.0]
+    
+    // MARK: - View Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -34,33 +36,11 @@ class PercentagesViewController : UIViewController, UITableViewDataSource, UITab
         }
     }
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        let cell: PercentageTableCell = tableView.dequeueReusableCell(withIdentifier: "percentageCell") as! PercentageTableCell
-        
-        let currentRowPercentage = percentages[indexPath.row]
-        
-        if units == MeasurementUnit.kilograms {
-            let percentage : Int = Int(currentRowPercentage * oneRepMaxWeightInKgs)
-            cell.weightTextLabel.text = "\(percentage) kg"
-        }
-        else {
-            let percentage : Int = Int(currentRowPercentage * oneRepMaxWeightInKgs * POUNDS_TO_KILOS_RATE)
-            cell.weightTextLabel.text = "\(percentage) lbs"
-        }
-      
-        cell.percentageTextLabel.text = "\(currentRowPercentage * 100.0)% of 1RM"
-        
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return percentages.count
-    }
-    
     override var preferredStatusBarStyle : UIStatusBarStyle {
         return UIStatusBarStyle.lightContent
     }
+    
+    // MARK: - Outlets
 
     @IBAction func unitsToggleChanged(_ sender: UISegmentedControl) {
         if sender.selectedSegmentIndex == 0 {
